Various refactorings
Move all torch modules into their own files.

Add PyRight config to turn on basic type checking; many files are now
red, and PyTorch seems to be extra difficult to type properly :(

Add some typing for code related to argparse.

Declare potentially undefined variables and add some typings where
PyRight complains.

Reword some log messages.
