+++ BertForDeprel/parser/modules/MLP.py
@@ -1,0 +1,28 @@
+from torch import nn
+from math import prod
+
+
+class MLP(nn.Module):
+    """Multi-layer perceptron with dropout."""
+    def __init__(self, input_size, layer_size, depth, activation, dropout):
+        super(MLP, self).__init__()
+        self.layers = nn.Sequential()
+        act_fn = getattr(nn, activation)
+        for i in range(depth):
+            self.layers.add_module('fc_{}'.format(i),
+                                   nn.Linear(input_size, layer_size))
+            if activation:
+                self.layers.add_module('{}_{}'.format(activation, i),
+                                       act_fn())
+            if dropout:
+                self.layers.add_module('dropout_{}'.format(i),
+                                       nn.Dropout(dropout))
+            input_size = layer_size
+
+    def forward(self, x):
+        return self.layers(x)
+
+    @property
+    def num_parameters(self):
+        """Returns the number of trainable parameters of the model."""
+        return sum(prod(p.shape) for p in self.parameters() if p.requires_grad)
