--- BertForDeprel/parser/utils/modules_utils.py
@@ -1,276 +1,0 @@
-import torch
-from torch import nn
-
-from torch.nn.modules.rnn import apply_permutation
-from torch.nn.utils.rnn import PackedSequence
-
-
-class MLP(nn.Module):
-    """Module for an MLP with dropout."""
-    def __init__(self, input_size, layer_size, depth, activation, dropout):
-        super(MLP, self).__init__()
-        self.layers = nn.Sequential()
-        act_fn = getattr(nn, activation)
-        for i in range(depth):
-            self.layers.add_module('fc_{}'.format(i),
-                                   nn.Linear(input_size, layer_size))
-            if activation:
-                self.layers.add_module('{}_{}'.format(activation, i),
-                                       act_fn())
-            if dropout:
-                self.layers.add_module('dropout_{}'.format(i),
-                                       nn.Dropout(dropout))
-            input_size = layer_size
-
-    def forward(self, x):
-        return self.layers(x)
-
-    @property
-    def num_parameters(self):
-        """Returns the number of trainable parameters of the model."""
-        return sum(prod(p.shape) for p in self.parameters() if p.requires_grad)
-
-
-class BiAffine(nn.Module):
-    """Biaffine attention layer."""
-    def __init__(self, input_dim, output_dim, bias_x=False, bias_y=False):
-        super(BiAffine, self).__init__()
-        self.bias_x = bias_x
-        self.bias_y = bias_y
-
-        self.input_dim = input_dim
-        self.output_dim = output_dim
-        self.U = nn.Parameter(torch.FloatTensor(output_dim, input_dim + self.bias_x, input_dim + self.bias_y))
-        nn.init.xavier_uniform_(self.U)
-
-
-    def forward(self, Rh, Rd):
-        if self.bias_x:
-            Rh = torch.cat((Rh, torch.ones_like(Rh[..., :1])), -1)
-
-        if self.bias_y:
-            Rd = torch.cat((Rd, torch.ones_like(Rd[..., :1])), -1)
-
-        Rh = Rh.unsqueeze(1)
-        Rd = Rd.unsqueeze(1)
-
-        S = Rh @ self.U @ Rd.transpose(-1, -2)
-
-        return S.squeeze(1)
-
-
-    @property
-    def num_parameters(self):
-        """Returns the number of trainable parameters of the model."""
-        return sum(prod(p.shape) for p in self.parameters() if p.requires_grad)
-
-
-
-class BiAffine2(nn.Module):
-
-    def __init__(self, n_in, n_out=1, bias_x=True, bias_y=True):
-        super(BiAffine2, self).__init__()
-
-        self.n_in = n_in
-        self.n_out = n_out
-        self.bias_x = bias_x
-        self.bias_y = bias_y
-        self.weight = nn.Parameter(torch.Tensor(n_out,
-                                                n_in + bias_x,
-                                                n_in + bias_y))
-        self.reset_parameters()
-
-    def extra_repr(self):
-        s = f"n_in={self.n_in}, n_out={self.n_out}"
-        if self.bias_x:
-            s += f", bias_x={self.bias_x}"
-        if self.bias_y:
-            s += f", bias_y={self.bias_y}"
-
-        return s
-
-    def reset_parameters(self):
-        nn.init.zeros_(self.weight)
-
-    def forward(self, x, y):
-        if self.bias_x:
-            x = torch.cat((x, torch.ones_like(x[..., :1])), -1)
-        if self.bias_y:
-            y = torch.cat((y, torch.ones_like(y[..., :1])), -1)
-        # [batch_size, n_out, seq_len, seq_len]
-        s = torch.einsum('bxi,oij,byj->boxy', x, self.weight, y)
-
-        # print("x", x.size())
-        # print("y", y.size())
-        # s = x @ self.weight @ y.transpose(-1, -2)
-        # print("s", s.size())
-        # remove dim 1 if n_out == 1
-        # s = s.unsqueeze(0)
-        s = s.squeeze(1)
-
-        return s
-
-
-
-
-class BiLSTM(nn.Module):
-
-    def __init__(self, input_size, hidden_size, num_layers=1, dropout=0):
-        super(BiLSTM, self).__init__()
-
-        self.input_size = input_size
-        self.hidden_size = hidden_size
-        self.num_layers = num_layers
-        self.dropout = dropout
-
-        self.f_cells = nn.ModuleList()
-        self.b_cells = nn.ModuleList()
-        for _ in range(self.num_layers):
-            self.f_cells.append(nn.LSTMCell(input_size=input_size,
-                                            hidden_size=hidden_size))
-            self.b_cells.append(nn.LSTMCell(input_size=input_size,
-                                            hidden_size=hidden_size))
-            input_size = hidden_size * 2
-
-        self.reset_parameters()
-
-    def __repr__(self):
-        s = self.__class__.__name__ + '('
-        s += f"{self.input_size}, {self.hidden_size}"
-        if self.num_layers > 1:
-            s += f", num_layers={self.num_layers}"
-        if self.dropout > 0:
-            s += f", dropout={self.dropout}"
-        s += ')'
-
-        return s
-
-    def reset_parameters(self):
-        for param in self.parameters():
-            # apply orthogonal_ to weight
-            if len(param.shape) > 1:
-                nn.init.orthogonal_(param)
-            # apply zeros_ to bias
-            else:
-                nn.init.zeros_(param)
-
-    def permute_hidden(self, hx, permutation):
-        if permutation is None:
-            return hx
-        h = apply_permutation(hx[0], permutation)
-        c = apply_permutation(hx[1], permutation)
-
-        return h, c
-
-    def layer_forward(self, x, hx, cell, batch_sizes, reverse=False):
-        hx_0 = hx_i = hx
-        hx_n, output = [], []
-        steps = reversed(range(len(x))) if reverse else range(len(x))
-        # if self.training:
-        #     hid_mask = SharedDropout.get_mask(hx_0[0], self.dropout)
-
-        for t in steps:
-            last_batch_size, batch_size = len(hx_i[0]), batch_sizes[t]
-            if last_batch_size < batch_size:
-                hx_i = [torch.cat((h, ih[last_batch_size:batch_size]))
-                        for h, ih in zip(hx_i, hx_0)]
-            else:
-                hx_n.append([h[batch_size:] for h in hx_i])
-                hx_i = [h[:batch_size] for h in hx_i]
-            hx_i = [h for h in cell(x[t], hx_i)]
-            output.append(hx_i[0])
-            if self.training:
-                hx_i[0] = hx_i[0] * hid_mask[:batch_size]
-        if reverse:
-            hx_n = hx_i
-            output.reverse()
-        else:
-            hx_n.append(hx_i)
-            hx_n = [torch.cat(h) for h in zip(*reversed(hx_n))]
-        output = torch.cat(output)
-
-        return output, hx_n
-
-    def forward(self, sequence, hx=None):
-        x, batch_sizes = sequence.data, sequence.batch_sizes.tolist()
-        batch_size = batch_sizes[0]
-        h_n, c_n = [], []
-
-        if hx is None:
-            ih = x.new_zeros(self.num_layers * 2, batch_size, self.hidden_size)
-            h, c = ih, ih
-        else:
-            h, c = self.permute_hidden(hx, sequence.sorted_indices)
-        h = h.view(self.num_layers, 2, batch_size, self.hidden_size)
-        c = c.view(self.num_layers, 2, batch_size, self.hidden_size)
-
-        for i in range(self.num_layers):
-            x = torch.split(x, batch_sizes)
-            if self.training:
-                mask = SharedDropout.get_mask(x[0], self.dropout)
-                x = [i * mask[:len(i)] for i in x]
-            x_f, (h_f, c_f) = self.layer_forward(x=x,
-                                                 hx=(h[i, 0], c[i, 0]),
-                                                 cell=self.f_cells[i],
-                                                 batch_sizes=batch_sizes)
-            x_b, (h_b, c_b) = self.layer_forward(x=x,
-                                                 hx=(h[i, 1], c[i, 1]),
-                                                 cell=self.b_cells[i],
-                                                 batch_sizes=batch_sizes,
-                                                 reverse=True)
-            x = torch.cat((x_f, x_b), -1)
-            h_n.append(torch.stack((h_f, h_b)))
-            c_n.append(torch.stack((c_f, c_b)))
-        x = PackedSequence(x,
-                           sequence.batch_sizes,
-                           sequence.sorted_indices,
-                           sequence.unsorted_indices)
-        hx = torch.cat(h_n, 0), torch.cat(c_n, 0)
-        hx = self.permute_hidden(hx, sequence.unsorted_indices)
-
-        return x, hx
-
-
-
-
-class BiAffineTrankit(nn.Module):
-    '''
-    implemented based on the paper https://arxiv.org/abs/1611.01734
-    '''
-
-    def __init__(self, in_dim1, in_dim2, hidden_dim, output_dim):
-        super().__init__()
-        self.in_dim1 = in_dim1
-        self.in_dim2 = in_dim2
-        self.hidden_dim = hidden_dim
-        self.output_dim = output_dim
-
-        self.ffn1 = nn.Sequential(
-            nn.Linear(in_dim1, hidden_dim),
-            nn.ReLU(),
-            nn.Dropout(0.5)
-        )
-        self.ffn2 = nn.Sequential(
-            nn.Linear(in_dim2, hidden_dim),
-            nn.ReLU(),
-            nn.Dropout(0.5)
-        )
-        # pairwise interactions
-        self.pairwise_weight = nn.Parameter(torch.Tensor(in_dim1 + 1, in_dim2 + 1, output_dim))
-        self.pairwise_weight.data.zero_()
-
-    def forward(self, x1, x2):
-        h1 = self.ffn1(x1)
-        h2 = self.ffn2(x2)
-        # make interactions
-        g1 = torch.cat([h1, h1.new_ones(*h1.size()[:-1], 1)], len(h1.size()) - 1)
-        g2 = torch.cat([h2, h2.new_ones(*h2.size()[:-1], 1)], len(h2.size()) - 1)
-
-        g1_size = g1.size()
-        g2_size = g2.size()
-
-        g1_w = torch.mm(g1.view(-1, g1_size[-1]), self.pairwise_weight.view(-1, (self.in_dim2 + 1) * self.output_dim))
-        g2 = g2.transpose(1, 2)
-        g1_w_g2 = g1_w.view(g1_size[0], g1_size[1] * self.output_dim, g2_size[2]).bmm(g2)
-        g1_w_g2 = g1_w_g2.view(g1_size[0], g1_size[1], self.output_dim, g2_size[1]).transpose(2, 3)
-        return g1_w_g2
