+++ pyproject.toml
@@ -1,0 +1,2 @@
+[tool.pyright]
+typeCheckingMode = "basic"
